// src/app/modules/lobby/models/view/private-table-list.ts
export type PrivateTable = PvtTables[];

export interface PvtTables {
  gv: string;
  ct: string;
  mxp: number;
  mip: number;
  bl: number;
  bh: number;
  sb: number;
  bb: number;
  wa: number;
  sd: string;
  ed: string;
  bict: number;
  cot: boolean;
  dpt: number;
  drmm: number;
  gt2: number;
  gt2l: number;
  cr: number;
  rhu: number;
  rc1: number;
  rc2: number;
  tba: boolean;
  tb: number;
  tbrr: number;
  tt: number;
  et: number;
  pnr: string;
  tbl: string;
  blbh: string;
  sbbb: string;
  prize: string;
  sts: string;
  own: string;
  jnd: string;
  rv: string;
  configId: number;
  blD: number;
  bhD: number;
  sbD: number;
  bbD: number;
  isSelected: boolean;
}